Implementa el componente ListadoHome, ademas limpieza
